pyproject.toml
[project]
name = "cosmochatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cryptography>=44.0.3",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "langchain>=0.3.25",
    "langchain-community>=0.3.23",
    "langchain-huggingface>=0.2.0",
    "langchain-milvus>=0.1.10",
    "langchain-openai>=0.3.16",
    "openai>=1.77.0",
    "openai-agents>=0.0.14",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "pymysql>=1.1.1",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "sentence-transformers>=4.1.0",
    "sqlmodel>=0.0.24",
    "tenacity>=9.1.2",
    "uvicorn>=0.34.2",
]
